:bug: show date range of pending period
