+++ lib/getRewards.js
@@ -51,8 +51,7 @@ const getSecondsInEpoch = () => {
 
 export const getEpochStart = (epoch) => {
   const start = parseInt(process.env.NEXT_PUBLIC_EPOCH_LAUNCH_UNIX_TIME)
-  const secondsInEpoch = getSecondsInEpoch()
-  return start + secondsInEpoch * (epoch - 1)
+  return start + getSecondsInEpoch() * (epoch - 1)
 }
 
 export const getEpochEnd = (epoch) => {
